Allow spaces (and words) after name tag
Any alphanumeric characters after the name tag were causing an opaque error (see Issue #75).  This happened because the parser allows for 0 or more of anything after the initial name string.  This element appears as another leaf in the parse tree, along with the name.

The error was thrown because the subsequent three element vector (the :name keyword, the name string, and the extra alphanumeric characters) was being passed to a `(merge {} ...)`. Adding a transformation that merged all elements under the name branch of the parse tree into on (trimmed) string fixed the error.